PCEM Conversion (#720)
There was an error for the conversion of point charge coordinates in external embedding. The Orca flag was wrongly set to "true" without an "interface=orca" command. This lead to deviations between point charge gradients calculated with the xTB binaries and our test suite. 
This will fix this. Will also close Issue #683 and PRs #718 and #719.